Code_Generator/requirements.txt
aiohttp==3.9.1; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
altair==5.2.0; python_version >= '3.8'
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.2.0; python_version >= '3.8'
async-timeout==4.0.3; python_version < '3.11'
attrs==23.2.0; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
cachetools==5.3.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
dataclasses-json==0.6.3; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
exceptiongroup==1.2.0; python_version < '3.11'
fastapi==0.108.0; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.40; python_version >= '3.7'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.2; python_version >= '3.8'
httpx==0.26.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
importlib-metadata==6.11.0; python_version >= '3.8'
jinja2==3.1.2; python_version >= '3.7'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==2.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonschema==4.20.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
langchain==0.0.354; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-community==0.0.8; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.1.6; python_version < '4.0' and python_full_version >= '3.8.1'
langsmith==0.0.77; python_version < '4.0' and python_full_version >= '3.8.1'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
marshmallow==3.20.1; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
multidict==6.0.4; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.3; python_version >= '3.9'
openai==1.6.1; python_full_version >= '3.7.1'
packaging==23.2; python_version >= '3.7'
pandas==2.1.4; python_version >= '3.9'
pillow==10.2.0; python_version >= '3.8'
protobuf==4.25.1; python_version >= '3.8'
pyarrow==14.0.2; python_version >= '3.8'
pydantic==2.5.3; python_version >= '3.7'
pydantic-core==2.14.6; python_version >= '3.7'
pydeck==0.8.1b0; python_version >= '3.7'
pygments==2.17.2; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.32.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rich==13.7.0; python_full_version >= '3.7.0'
rpds-py==0.16.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.25; python_version >= '3.7'
starlette==0.32.0.post1; python_version >= '3.8'
streamlit==1.29.0; python_version >= '3.8' and python_full_version != '3.9.7'
tenacity==8.2.3; python_version >= '3.7'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.12.0; python_version >= '3.5'
tornado==6.4; python_version >= '3.8'
tqdm==4.66.1; python_version >= '3.7'
typing-extensions==4.9.0; python_version < '3.11'
typing-inspect==0.9.0
tzdata==2023.4; python_version >= '2'
tzlocal==5.2; python_version >= '3.8'
urllib3==2.1.0; python_version >= '3.8'
uvicorn==0.25.0; python_version >= '3.8'
validators==0.22.0; python_version >= '3.8'
watchdog==3.0.0; platform_system != 'Darwin'
yarl==1.9.4; python_version >= '3.7'
zipp==3.17.0; python_version >= '3.8'
